requirements.txt
discord.py[voice]
googletrans
pillow
gtts
requests
yahoo_fin
pandas